@@ -4,11 +4,10 @@
     "homepage": "https://codegreencreative.com/",
     "license": "MIT",
     "require": {
-        "php": ">=5.3.3",
+        "php": "^7.1.3",
         "ext-json": "*",
         "ext-hash": "*",
-        "illuminate/support": ">=4.1.0",
-        "illuminate/routing": ">=4.1.0"
+        "laravel/framework": ">=5.7"
     },
     "autoload": {
         "psr-4": {
