@@ -2,10 +2,8 @@
 
 namespace CodeGreenCreative\Aweber;
 
-use Cache;
 use CodeGreenCreative\Aweber\Aweber\AweberApi;
-use CodeGreenCreative\Aweber\Exceptions\AweberException;
-use Config;
+use CodeGreenCreative\Aweber\Aweber\Exceptions\AweberException;
 
 class AweberClient
 {
@@ -30,11 +28,11 @@ class AweberClient
         $this->api_url = 'https://api.aweber.com/1.0';
         $this->oauth_url = 'https://auth.aweber.com/1.0/oauth';
 
-        $this->store = Config::get('laravel-aweber::cache', Config::get('aweber.cache', null));
-        $this->client_id = Config::get('laravel-aweber::client_id', Config::get('aweber.client_id', null));
-        $this->client_secret = Config::get('laravel-aweber::client_secret', Config::get('aweber.client_secret', null));
-        $this->token_key = Config::get('laravel-aweber::token_key', Config::get('aweber.token_key', null));
-        $this->token_secret = Config::get('laravel-aweber::token_secret', Config::get('aweber.token_secret', null));
+        $this->store = config('laravel-aweber::cache', config('aweber.cache', null));
+        $this->client_id = config('laravel-aweber::client_id', config('aweber.client_id', null));
+        $this->client_secret = config('laravel-aweber::client_secret', config('aweber.client_secret', null));
+        $this->token_key = config('laravel-aweber::token_key', config('aweber.token_key', null));
+        $this->token_secret = config('laravel-aweber::token_secret', config('aweber.token_secret', null));
 
         if (empty($this->client_id)) {
             throw new AweberException('Client ID is not set');
@@ -42,12 +40,29 @@ class AweberClient
         if (empty($this->client_secret)) {
             throw new AweberException('Client secret is not set');
         }
-        $this->aweber_api = new AweberApi($this->client_id, $this->client_secret);
-        $aweber_account = $this->aweber_api->getAccount($this->token_key, $this->token_secret);
-        $this->account_id = $aweber_account->data['id'];
 
+        $this->aweber_api = new AweberApi($this->client_id, $this->client_secret);
+
+        if ($this->token_key && $this->token_secret) {
+            $this->setConsumer($this->token_key, $this->token_secret);
+        }
+    }
+
+    /**
+     * Set up the HTTP client
+     *
+     * @param string $token
+     * @param string $secret
+     * @return self
+     */
+    public function setConsumer(string $token, string $secret): self
+    {
+        $aweber_account = $this->aweber_api->getAccount($token, $secret);
+        $this->account_id = $aweber_account->data['id'];
         // Transform base URI
         $this->base_uri = $this->api_url . $aweber_account->url . '/';
+
+        return $this;
     }
 
     /**
@@ -61,10 +76,9 @@ class AweberClient
      * @param  array  $data
      * @return integer | array
      */
-    public function request($method, $path, $data = array(), $options = array())
+    public function request($method, $path, $data = [], $options = [])
     {
         $uri = $this->base_uri . $path;
-
         if (config('aweber.allow_empty')) {
             $options = array_merge($options, ['allow_empty' => true]);
         }
@@ -132,10 +146,10 @@ class AweberClient
         $handle = curl_init($this->api_url . 'accounts');
         curl_setopt($handle, CURLOPT_RETURNTRANSFER, true);
         curl_setopt($handle, CURLOPT_HEADER, false);
-        curl_setopt($handle, CURLOPT_HTTPHEADER, array(
+        curl_setopt($handle, CURLOPT_HTTPHEADER, [
             'Content-Type: application/json',
-            'Authorization: Bearer ' . $this->token->access_token
-        ));
+            'Authorization: Bearer ' . $this->token->access_token,
+        ]);
         curl_setopt($handle, CURLOPT_SSL_VERIFYPEER, false);
         curl_setopt($handle, CURLOPT_VERBOSE, false);
         curl_setopt($handle, CURLOPT_CONNECTTIMEOUT, 10);
